requirements.txt
certifi==2018.10.15
chardet==3.0.4
cookies==2.2.1
Django==1.11.16
django-constance==2.3.1
django-picklefield==1.1.0
djangorestframework==3.9.0
funcsigs==1.0.2
idna==2.7
mock==2.0.0
pbr==5.1.1
pytz==2018.7
requests==2.20.1
responses==0.10.4
six==1.11.0
urllib3==1.24.1
